Did some renaming of classes, and is sounding better
