+++ www/configurator/src/app/iframe-manager/iframe-manager.spec.js
